move fetch-cache-limit test to app-static test
