+++ test/e2e/app-dir/app-static/app/force-cache/large-data/page.js
@@ -1,0 +1,13 @@
+export default async function Page() {
+  const resp = await fetch(
+    `http://localhost:${process.env.PORT}/api/large-data`,
+    {
+      cache: 'force-cache',
+    }
+  )
+  const data = await resp.json()
+
+  return <p id="content">{data.content}</p>
+}
+
+export const dynamic = 'force-dynamic'
