--- test/e2e/fetch-cache-limit/app/layout.tsx
@@ -1,14 +1,0 @@
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  )
-}
-
-export const dynamic = 'force-dynamic'
