--- test/e2e/fetch-cache-limit/fetch-cache-limit.test.ts
@@ -1,22 +1,0 @@
-import { createNextDescribe } from 'e2e-utils'
-
-createNextDescribe(
-  'fetch-cache-limit',
-  {
-    files: __dirname,
-    // don't have access to runtime logs on deploy
-    skipDeployment: true,
-  },
-  ({ next }) => {
-    it('when awaiting `fetch` using an unknown domain, stack traces are preserved', async () => {
-      const browser = await next.browser('/')
-      await browser.waitForElementByCss('#content')
-      await browser.refresh()
-      await browser.waitForElementByCss('#content')
-      await browser.refresh()
-      await browser.waitForElementByCss('#content')
-      // console.log(next.cliOutput.match('Load data'))
-      expect(next.cliOutput.match('Load data').length).toBe(1)
-    })
-  }
-)
