--- test/e2e/fetch-cache-limit/next.config.js
@@ -1,6 +1,0 @@
-/** @type {import('next').NextConfig} */
-module.exports = {
-  experimental: {
-    incrementalCacheHandlerPath: require.resolve('./cache-handler.js'),
-  },
-}
